+++ Crafttweaker_Annotation_Processors/src/main/java/com/blamejared/crafttweaker_annotation_processors/processors/document/shared/util/CommentUtils.java
@@ -1,6 +1,5 @@
-package com.blamejared.crafttweaker_annotation_processors.processors.document.shared;
+package com.blamejared.crafttweaker_annotation_processors.processors.document.shared.util;
 
-import com.blamejared.crafttweaker_annotation_processors.processors.StringReplaceUtil;
 import com.blamejared.crafttweaker_annotation_processors.processors.document.shared.types.DocumentedType;
 
 import javax.annotation.processing.ProcessingEnvironment;
